+++ src/main/java/com/metamx/metrics/cgroups/CgroupDiscoverer.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.metamx.metrics.cgroups;
 
 import java.nio.file.Path;
