+++ src/main/java/com/metamx/metrics/cgroups/JvmPidDiscoverer.java
@@ -1,0 +1,57 @@
+/*
+ * Copyright 2017 Metamarkets Group Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.metamx.metrics.cgroups;
+
+import com.metamx.common.RE;
+import java.lang.management.ManagementFactory;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
+import java.util.regex.Pattern;
+
+public class JvmPidDiscoverer implements PidDiscoverer
+{
+  private static final String KEY = "pid";
+  private final ConcurrentMap<String, Long> pid = new ConcurrentHashMap<>();
+
+
+  /**
+   * Returns the PID as a best guess. This uses methods that are not guaranteed to actually be the PID.
+   * <p>
+   * TODO: switch to ProcessHandle.current().getPid() for java9 potentially
+   *
+   * @return the PID of the current jvm if available
+   *
+   * @throws RuntimeException if the pid cannot be determined
+   */
+  @Override
+  public long getPid()
+  {
+    return pid.computeIfAbsent(KEY, unused -> {
+      final String jvmName = ManagementFactory.getRuntimeMXBean().getName();
+      final String[] nameSplits = jvmName.split(Pattern.quote("@"));
+      if (nameSplits.length != 2) {
+        throw new RE("Unable to determine pid from [%s]", jvmName);
+      }
+      try {
+        return Long.parseLong(nameSplits[0]);
+      }
+      catch (NumberFormatException nfe) {
+        throw new RE(nfe, "Unable to determine pid from [%s]", jvmName);
+      }
+    });
+  }
+}
