--- src/test/java/com/metamx/metrics/cgroups/InterningJvmPidDiscovererTest.java
@@ -1,28 +1,0 @@
-/*
- * Copyright 2017 Metamarkets Group Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.metamx.metrics.cgroups;
-
-import org.junit.Assert;
-import org.junit.Test;
-
-public class InterningJvmPidDiscovererTest
-{
-  @Test
-  public void getPid() throws Exception
-  {
-    Assert.assertNotNull(new InterningJvmPidDiscoverer().getPid());
-  }
-}
