+++ src/test/java/com/metamx/metrics/cgroups/ProcCgroupDiscovererTest.java
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package com.metamx.metrics.cgroups;
 
 import com.google.common.collect.ImmutableSet;
